% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVictims.R
\name{addVictims}
\alias{addVictims}
\title{Add victim samples to a DVI dataset}
\usage{
addVictims(dvi, newvics)
}
\arguments{
\item{dvi}{A \code{\link[=dviData]{dviData()}} object.}

\item{newvics}{A \code{singleton} object or a list of such.}
}
\value{
A new \code{\link[=dviData]{dviData()}} object with the new victims added to the \code{pm}
data.
}
\description{
This is a convenience function for adding new victim samples to an existing
\code{\link[=dviData]{dviData()}} object. Users are recommended to use this instead of manipulating
the \code{pm} data directly, as it ensures consistency of the output and also
recomputes the internal \code{pairings} object.
}
\examples{
addVictims(example1, singleton("V4"))

}
