% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkInput}
\alias{checkInput}
\title{Checks input}
\usage{
checkInput(pm, am, vp, mp)
}
\arguments{
\item{pm}{A list of singeltons}

\item{am}{A list of pedigrees}

\item{vp}{Character vector with names of victims}

\item{mp}{Character vector with names of missing persons}
}
\value{
list with NULL if OK otherwise error message and null likelihood
}
\description{
Checks input
}
\examples{
vp = c("V1", "V2")
pm = singletonList(2, ids = vp)
m = marker(pm[[1]], alleles = 1:2, "V1" = 1:2)
pm[[1]] = addMarkers(pm[[1]], m)
m = marker(pm[[2]], alleles = 1:2, "V2" = 1:2)
pm[[2]] = addMarkers(pm[[2]], m)
mp = c("MP1", "MP2")
am = nuclearPed(2, children = mp, father = "R1", mother = "R2")
m = marker(am, "R1" = 1:2, "R2"= 1:2)
am = addMarkers(am, m)
checkInput(pm, am, vp, mp)
checkInput(pm, am, vp[c(1,1)], mp)
checkInput(pm, am, vp, mp[c(1,1)])
m = marker(am, "R1" = 1:2, "R2"= 1:2, "MP1" = 1)
am = addMarkers(am, m)
checkInput(pm, am, vp, mp) 
am = setAlleles(am, alleles = 0)
m = marker(am, "R1" = 1:2, "R2"= 1:2)
am = addMarkers(am, m)
pm = setAlleles(pm[[1]], alleles = 0)
checkInput(pm, am, vp, mp)
}
