% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataExample483}
\alias{dataExample483}
\title{Data. Exercise, Kling et al. (2021)}
\format{
A list of 3 elements:

  * `pm`: A list of 5 singletons (victims).

  * `am`: A list of 1 pedigree 

  * `missing`: A vector containing the names of the missing persons.
}
\usage{
dataExample483
}
\description{
Data used in Example 4.8.3 in Kling et al. (2021) 
"Mass Identifications: Statistical Methods in Forensic Genetics"
There are victims V1 (female), V2 (male), V3 (female), V4 (male) and V5 (female). 
There is one reference family with missing persons
12 missing persons, 6 females and 6 males.
There are 13 codis markers, equal mutation model, rate 0.001.
}
\examples{
\donttest{
pm = dataExample483$pm
am = dataExample483$am
missing = dataExample483$missing
ncomb(3,2,6,6)
# Make large window prior to plotting, e.g. by windows()
plotPedList(list(pm, am), marker = 1:2, hatched = typedMembers, 
            col = list(red = missing))
res = jointDVI(pm, am, missing)
# Transfer victims and check solution
amId = transferMarkers(pm, am, idsFrom = paste0("V",1:5), 
                       idsTo = missing[1:5], erase = FALSE)
plot(amId, marker = 1:2, hatched = typedMembers, 
            col = list(red = missing))
}
}
\keyword{datasets}
