% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataExercise498}
\alias{dataExercise498}
\title{Data used in the book Kling et al. (2021)}
\format{
A list of 3 elements:

  * `pm`: A list of 3 singletons (victims).

  * `am`: A list of 1 pedigree. 

  * `missing`: A vector containing the names of the missing persons.
}
\usage{
dataExercise498
}
\description{
Data used in Exercise 4.9.8 in Kling et al. (2021) 
"Mass Identifications: Statistical Methods in Forensic Genetics".
There are 2 female victims and 1 male. 
There is one reference family with 2 missing females and one missing male. 
There are 16 markers, equal mutation model, rate 0.001.
}
\keyword{datasets}
