% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviJoint.R
\name{denominatorRow}
\alias{denominatorRow}
\title{Row number of denominator assignment}
\usage{
denominatorRow(assignments)
}
\arguments{
\item{assignments}{A data frame.}
}
\value{
A matrix of the same dim as \code{assignments}.
}
\description{
This is an auxiliary function used in the calculation of conditional LRs from
a table of joint LRs. For each assignment vector (a row in the table) and
each nontrivial pairing \code{V = M} within that assignment, it find the row
containing the assignment to be used in the denominator, i.e., where \code{V = M}
is replaced with the nonpairing \verb{V = *}. Table entries of \code{*}, or where
the denominator does not occur, gives NA.
}
\examples{
\dontrun{
a = rbind(c(V1 = "M1", V2 = "M2", V3 = "M3"),
          c(V1 = "M1", V2 = "M2", V3 = "*"),
          c(V1 = "*",  V2 = "M2", V3 = "*"))
a
dvir:::denominatorRow(a)
}
}
\keyword{internal}
