% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvi.3.3.R
\docType{data}
\name{dvi.3.3}
\alias{dvi.3.3}
\title{Data for dvir: Example 3.3 in EKM}
\description{
A list with 4 elements: "pm" (post mortem), "vict" (names of 8 victims),
"am" (ante mortem), "miss" (names of  5 missing persons). 
"pm" and "am" are ped objects
or lists of such while "vict" and "miss" are character vectors 
giving names of victims
and missing persons. There are 13 CODIS marker, with a proportional mutation
model with rate = 0.003. `dvi.3.3.nomut` same except for no mutation model.
}
\examples{

data(dvi.3.3)
}
\keyword{data}
