% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvi.nfi.R
\docType{data}
\name{dvi.nfi}
\alias{dvi.nfi}
\alias{dvi.nfi.mut}
\title{Data for dvir: ICMP example}
\description{
A list with 4 elements: "pm" (post mortem), "vict" (names of 5 victims),
"am" (ante mortem), "miss" (names of  12 missing persons). 
"pm" and "am" are ped objects
or lists of such while "vict" and "miss" are character vectors 
giving names of victims
and missing persons. `dvi.nfi.mut` same except fror a `proportional`mutation model
with rate 0.005.
}
\examples{

data(dvi.nfi)
}
\keyword{data}
