% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvir-package.R
\docType{package}
\name{dvir-package}
\alias{dvir}
\alias{dvir-package}
\title{dvir: Disaster Victim Identification}
\description{
Joint DNA-based disaster victim identification (DVI), as described in Vigeland and Egeland (2021) \doi{10.21203/rs.3.rs-296414/v1}. Identification is performed by optimising the joint likelihood of all victim samples and reference individuals. Individual identification probabilities, conditional on all available information, are derived from the joint solution in the form of posterior pairing probabilities. 'dvir' is part of the 'pedsuite' collection of packages for pedigree analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/dvir}
  \item Report bugs at \url{https://github.com/magnusdv/dvir/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

Authors:
\itemize{
  \item Thore Egeland \email{thore.egeland@nmbu.no} (\href{https://orcid.org/0000-0002-3465-8885}{ORCID})
}

}
\keyword{internal}
