% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.3.3.R
\name{example.3.3}
\alias{example.3.3}
\title{Exercise 3.3 of EKM}
\usage{
example.3.3(from, to, ids.from, ids.to, simID, moveID, N = 1,
  limit = 1, seed = NULL, mutRateSim = 0, mutRateCalc = 0)
}
\arguments{
\item{from}{A list of singeltons}

\item{to}{A list of pedigrees}

\item{ids.from}{Character vector with names of victims}

\item{ids.to}{Character vector with names of missing persons}

\item{simID}{A list of mp-s to simulate}

\item{moveID}{A list of victims to be replaced by the simulkated persons}

\item{N}{integer No of simulations}

\item{limit}{Double}

\item{seed}{Integer}

\item{mutRateSim}{double}

\item{mutRateCalc}{double}
}
\description{
The exercises 3.3 of EKM is studied vie simulation
}
\details{
The simulated results are evaluated by the simple marginal search which
is sufficient for this example because of the informative realtives
}
\examples{

data(dvi.3.3.nomut) # Without mutations
dat = dvi.3.3.nomut
from = dat$pm
to = dat$am
ids.from = dat$vict
ids.to =  dat$miss
simID = ids.to
moveID = ids.from[1:5]
x1 = example.3.3(from, to, ids.from, ids.to, simID, moveID, 
N = 100, seed = 123, limit = 1,mutRateSim = 0.003, mutRateCalc = 0)

x2 = example.3.3(from, to, ids.from, ids.to, simID, moveID, 
N = 100, seed = 123, limit = 1,mutRateSim = 0.003, mutRateCalc = 0.003)

Plot
windows()
plotPedList(c(to), newdev = FALSE, marker=1, skip.empty.genotypes = TRUE, frames = T, 
frametitles = paste("F",1:5, sep="") )
plotPedList(c(from), newdev = FALSE, marker=1, 
skip.empty.genotypes = TRUE,  frames = 1:8)
}
