% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{Data. Stepwise identification may fail}
\format{
A list of 3 elements:
\itemize{
\item \code{pm}: A list of 3 singletons (victims).
\item \code{am}: A pedigree with three missing persons and one typed reference
individual.
\item \code{missing}: A vector containing the names of the missing persons.
}
}
\usage{
example1
}
\description{
A proof-of-concept dataset showing that showing that stepwise victim
identification may fail
}
\keyword{datasets}
