% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2}
\alias{example2}
\title{DVI dataset: Two reference families}
\format{
A list of 3 elements:
\itemize{
\item \code{pm}: A list of 3 singletons (victims).
\item \code{am}: A list of 2 pedigrees with three missing persons and one typed reference
individual.
\item \code{missing}: A vector containing the names of the missing persons.
}
}
\usage{
example2
}
\description{
A small DVI example with three victims, and three missing persons from two reference
families
}
\examples{
   
pm = example2$pm
am = example2$am
missing = example2$missing
jointDVI(pm, am, missing)
}
\keyword{datasets}
