% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise497.R
\name{exercise497}
\alias{exercise497}
\title{Solution Exercise 4.9.7 in the book Kling et al. (2021)}
\usage{
exercise497(
  pm,
  am,
  missing,
  nsim = 2,
  seed = NULL,
  simRef = TRUE,
  disableMutations = FALSE,
  undisputed = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pm}{A list of singletons.}

\item{am}{A list of pedigrees.}

\item{missing}{Character vector with names of missing persons.}

\item{nsim}{Number of simulations.}

\item{seed}{Integer.}

\item{simRef}{Logical. If TRUE, references are also simulated.}

\item{disableMutations}{Logical, see \code{\link[=jointDVI]{jointDVI()}}.}

\item{undisputed}{Logical, see \code{\link[=jointDVI]{jointDVI()}}.}

\item{verbose}{A logical.}
}
\value{
A list with two elements, the first the fraction of 'correct' solutions, the second
a matrix with first line from \code{\link[=jointDVI]{jointDVI()}}.
}
\description{
This is a DVI case with 3 female victims and 3 missing females
in three reference families. There are 23 markers with equal mutation rate 0.001.
Data are simulated from the solution V1 = MP1, V2 = MP2, V3 = MP3 and the purpose is
to check fraction of times the 'correct' solutions is obtained.
}
\seealso{
\code{\link[=jointDVI]{jointDVI()}}
}
