% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.nodup2.R
\name{expand.grid.nodup2}
\alias{expand.grid.nodup2}
\title{Combinations of victims without duplication}
\usage{
expand.grid.nodup2(lst, pm = NULL, am = NULL, sex = NULL, age = NULL)
}
\arguments{
\item{lst}{A list of vectors.}

\item{pm}{A list of singletons.}

\item{am}{A list of pedigrees.}

\item{sex}{A list of the sexes of the victims. (1=male,2=female,0=unknown)}

\item{age}{A list of age constraints for the victims, these are compared
against the structure of the pedigree.}
}
\value{
A data frame. Each row is a possible solution to the disaster victim
identification problem.
}
\description{
This is based on the function \code{\link[=expand.grid]{expand.grid()}}, but also removes combinations
with repeated elements. This function is specifically designed to deal with
DVI, and uses DVI data to further cut down on combinations. The function
takes sex and age as parameters, and removes impossible combinations based
on this information.
}
\examples{
library(stringr) # removed importfrom
pm = example2$pm
am = example2$am
missing = example2$missing

am[1] = setAlleles(am[1],"R1", alleles = 0)
jointRes = jointDVI(pm,am,missing,verbose = TRUE)

pairss = list(V1=c("*","M1","M2"),V2=c("*","M1","M2"),V3=c("M3","*"))
expand.grid.nodup2(pairss,pm,am,age="V2>V1")


}
\seealso{
\code{\link[=expand.grid]{expand.grid()}}
}
