% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{DVI dataset: Family of fire victims}
\format{
A \code{dviData} object with the following content:
\itemize{
\item \code{pm}: A list of 3 singletons (victims).
\item \code{am}: A pedigree with three missing persons and one typed reference
individual.
\item \code{missing}: A vector containing the names of the missing persons.
}
}
\usage{
fire
}
\description{
A family with three missing persons after a fire, and one reference
individual. This example is featured in the GLR paper (Egeland & Vigeland,
2024).
}
\examples{
fire

plotDVI(fire, marker = 1)

jointDVI(fire)

}
\keyword{datasets}
