% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freqsBlind}
\alias{freqsBlind}
\title{Allele frequencies}
\format{
A list of length 198. Each element is a vector of allele frequencies
  for a single marker. The first 27 markers are STRs, the remaining are SNPs.
}
\usage{
freqsBlind
}
\description{
A dataset containing allele frequencies for 198 markers.
}
\keyword{datasets}
