% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Find next candidates}
\usage{
generate(from, to, ids.from, ids.to, one = FALSE)
}
\arguments{
\item{from}{A list of singeltons}

\item{to}{A list of pedigrees}

\item{ids.from}{Character vector with names of victims}

\item{ids.to}{Character vector with names of missing persons}

\item{one}{logic If TRUE, only one-step}
}
\value{
list of matrices of one and two-moves
}
\description{
New moves are generated, NULL
if none are possible, for one and two step
}
\examples{
ids.from = c("V1", "V2")
from = singletonList(2, ids = ids.from)
m = marker(from[[1]], alleles = 1:2, "V1" = 1)
from[[1]] = addMarkers(from[[1]], m)
m = marker(from[[2]], alleles = 1:2, "V2" = 1:2)
from[[2]] = addMarkers(from[[2]], m)
ids.to = c("MP1", "MP2")
to = nuclearPed(2, children = ids.to, father = "R1", mother = "R2")
m = marker(to, "R1" = 1:2, "R2"= 1:2)
to = addMarkers(to, m)
# The thre possibilities: a move of two, one or none:
tab = generate(from, to, ids.from, ids.to, one = FALSE)
from[[1]] = swapSex(from[[1]], "V1")
tab = generate(from, to, ids.from,ids.to)
from[[2]] = swapSex(from[[2]], "V2")
tab = generate(from, to, ids.from,ids.to)
}
