% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviData.R
\name{getDatabase}
\alias{getDatabase}
\title{Extract the marker database from a DVI dataset}
\usage{
getDatabase(dvi, check = TRUE, includeAttrs = FALSE)
}
\arguments{
\item{dvi}{A \code{\link[=dviData]{dviData()}} object.}

\item{check}{A logical. If TRUE (default), the functions checks if the AM and
PM databases are identical, and raises an error if not. If FALSE, the AM
database is returned without checking PM.}

\item{includeAttrs}{A logical, indicating if the full list of attributes of
each marker (including mutation model and position data) should be
returned, or only the allele frequencies. Default: FALSE.}
}
\value{
A list of marker information.
}
\description{
Extract the marker database from a DVI dataset
}
\examples{
getDatabase(example1)
getDatabase(example1, includeAttrs = TRUE)

}
