% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviData.R
\name{getFamily}
\alias{getFamily}
\title{Get AM component of selected individuals}
\usage{
getFamily(dvi, ids)
}
\arguments{
\item{dvi}{A \code{\link[=dviData]{dviData()}} object.}

\item{ids}{A vector of ID labels of members of \code{dvi$am}.}
}
\value{
A vector of the same length as \code{ids}, containing the family names (if
\code{dvi$am} is named) or component indices (otherwise) of the \code{ids}
individuals.
}
\description{
Get AM component of selected individuals
}
\examples{
getFamily(example2, ids = example2$missing)

}
