% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs.R
\name{getIDs}
\alias{getIDs}
\title{Get IDs from list of singeltons or pedigrees}
\usage{
getIDs(peds)
}
\arguments{
\item{peds}{list of sigletons or pedigree}
}
\value{
Character vector with IDs
}
\description{
Get IDs from list of singeltons or pedigrees
}
\examples{
pm1 = singleton("PM1")
getIDs(pm1)
pm2 = singleton("PM2")
getIDs(list(pm1, pm2))
fam1 = nuclearPed(1, father = "R1", mother = "R2", child= "MP1" )
getIDs(fam1)
fam2 = nuclearPed(1, father = "R3", mother = "R4", child= "MP2" )
am = list(fam1, fam2)
getIDs(am)
}
