% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{global}
\alias{global}
\title{Search for best DVI solution}
\usage{
global(from, to, ids.to, moves, limit = 0, verbose = F)
}
\arguments{
\item{from}{A list of singletons.}

\item{to}{A list of pedigrees.}

\item{ids.to}{Character vector with names of missing persons.}

\item{moves}{List of length equal from with possible marginal moves.}

\item{limit}{Double. Only moves with LR above limit are kept.}

\item{verbose}{Logical.}
}
\value{
A data frame. Each row describes an a priori possible move. The log likelihood and 
the null likelihood is given as well as the LR with the null hypothesis in the numerator.
There is limited checking. For checking, preprocess with `marginal`.
}
\description{
Victims are given as a list of singletons and references as list of pedigrees.
Based on a specification of moves for each victim, if any, all possible moves 
are evaluated and a solution ranked according to the likelihood is returned.
}
\details{
This is currently a brute force approach, all possibilities are evaluated
}
\examples{
\dontrun{
library(forrel)
library(pedprobr)
n = 7
ids.from = paste("V", 1:n, sep = "")
sex = c(rep(1, n-1), 2)
df = data.frame(id = ids.from, fid = 0, mid = 0, sex = sex, 
                a1 = c(1,2,1,1,2,2,2), 
                a2 = c(1,2,1,1,2,2,2))
locus_annotations = list(alleles = 1:3, afreq = c(1, 1, 1)/3)
from = as.ped(df, locusAttributes = locus_annotations)
to = nuclearPed(3, father = "R1", mother = "R2", children= c("MP1","MP2","MP3"))
m = marker(to, "R1" = 1, "R2" = 1,   alleles = 1:3, afreq = c(1, 1, 1)/3, name = "a1")
to = addMarkers(to, m)
ids.to = c("MP1", "MP2","MP3")
plotPedList(list(from, to), marker = 1)
moves = list(V1 = c("V1", "MP1", "MP2"), V3 = c("MP1","MP2", "MP3"), V4 = c("V3", "MP3"), 
             V7 = c("V7","MP3"))
res = global(from, to, ids.to, moves, limit = 0)
moves = moves[1:2]
res = global(from, to, ids.to, moves, limit = 0)
moves = moves[1]
res = global(from, to, ids.to, moves, limit = 0)

# Example 2
# Checked against; http://familias.name/BookKEP/globalExample2.fam
Several reference families and relationship between victims
V1 = singleton("V1")
m = marker(name = "m1", V1, alleles = 1:10, "V1" = c(1,1))
V1 = setMarkers(V1, m)
V2 = singleton("V2")
m = marker(name = "m1", V2, alleles = 1:10, "V2" = c(1,2))
V2 = setMarkers(V2, m) 
V3 = singleton("V3",2)
m = marker(name = "m1", V3, alleles = 1:10, "V3" = c(3,4))
V3 = setMarkers(V3, m) 
V4 = singleton("V4")
m = marker(name = "m1", V4, alleles = 1:10, "V4" = c(3,4))
V4 = setMarkers(V4, m)  
from = list(V1 = V1, V2 = V2, V3 = V3, V4 = V4)
AM1 = nuclearPed(1, father = "MP1", child = "MP2", mother ="R1")
m = marker(name = "m1",AM1, alleles = 1:10, "R1" = c(2, 2))
AM1 = setMarkers(AM1, m)
AM2 = halfSibPed(sex1 = 1, sex2 = 2)
AM2 = relabel(AM2, c("R2", "MO2", "FA2", "MP3", "MP4"), 1:5)
m = marker(name = "m1", AM2, alleles = 1:10, "R2" = c(3,3))
AM2 = setMarkers(AM2, m)
to = list(AM1 = AM1, AM2 = AM2)
plotPedList(list(from, to), marker = "m1", shaded = typedMembers)
ids.to = c("MP1", "MP2", "MP3", "MP4") # user specified
males = ids.to[1:3]
females = ids.to[4]
moves = list(V1 = c("V1", males), 
             V2 = c("V2", males),
             V3 = c("V3", females), 
             V4 = c("V4", males))
res = global(from, to, ids.to, moves = moves, limit = 0, verbose = T)
resmarg = marginal(from, to, ids.to, moves = moves, limit = 0, 
             verbose = T, sorter = TRUE, nkeep = 2)
res = global(from, to, ids.to, moves = resmarg[[1]], limit = 0, verbose = T)

#From fam file
x = readFam("http://familias.name/BookKEP/globalExample2.fam")
from = x[[1]][1:4]
to = x[[1]][5:6]
plotPedList(list(from,to))
ids.to = c("MP1", "MP2", "MP3", "MP4") # user specified
males = ids.to[1:3]
females = ids.to[4]
moves = list(V1 = c("V1", males), V2 = c("V2", males),
             V3 = c("V3", females), V4 = c("V4", males))
res = global(from, to, ids.to, moves = moves, limit = 0, verbose = T)
}
}
\seealso{
`marginal`.
}
