% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heli}
\alias{heli}
\title{DVI dataset: Helicopter accident}
\format{
A \code{dviData} object with the following content:
\itemize{
\item \code{pm}: A list of 7 singletons (PM samples).
\item \code{am}: A list of 3 reference pedigrees.
\item \code{missing}: A vector containing the names of the 4 missing persons.
}
}
\usage{
heli
}
\description{
A DVI dataset based on Exercise 4.9.2 in Kling et al. (2021) "Mass Identifications".
The scenario concerns a helicopter accident with seven PM samples and three reference
families, containing four missing persons in total.
}
\details{
Two features are of particular interest. First, one of the families (F2) contains two
missing individuals. Second, several PM samples originate from the same individual,
necessitating a preparatory step merging these samples before matching against
the reference families.
}
\examples{
heli
plotDVI(heli)

# Identify and merge matching PM samples
m = mergePM(heli$pm, threshold = 1e6, dropout = 0.1)
names(m)

heli2 = setVictims(heli, m$pmReduced)
heli2

}
\keyword{datasets}
