% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KETPex481}
\alias{KETPex481}
\title{Data used in the book Kling et al. (2021)}
\format{
A \code{dviData} object with the following content:
\itemize{
\item \code{pm}: A list of 2 singletons (victims).
\item \code{am}: A list of 1 pedigree.
\item \code{missing}: A vector containing the names of the missing persons.
}
}
\usage{
KETPex481
}
\description{
Data used in Example 4.8.1 in Kling et al. (2021) "Mass Identifications:
Statistical Methods in Forensic Genetics". There  victims are V1 and V2, both
females. There is one reference family with 2 missing persons, both females.
There are 21 markers, no mutation model.
}
\examples{
plotDVI(KETPex481, marker = 1)

}
\keyword{datasets}
