% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik1.R
\name{lik1}
\alias{lik1}
\title{Move individuals, find likelihoods and update}
\usage{
lik1(from, to, pm, am)
}
\arguments{
\item{from}{Character vector with names of victims}

\item{to}{Character vector with names of missing persons}

\item{pm}{A list of singeltons}

\item{am}{A list of pedigrees}
}
\value{
Likelihood and updated pm and am lists
}
\description{
Individuals are moved from pm to am data.
The likelihood is calculated and pm and am
are updated
}
\examples{
vp = c("V1", "V2")
pm = singletonList(2, ids = vp)
m = marker(pm[[1]], alleles = 1:2, "V1" = 1:2)
pm[[1]] = addMarkers(pm[[1]], m)
m = marker(pm[[2]], alleles = 1:2, "V2" = 1:2)
pm[[2]] = addMarkers(pm[[2]], m)
mp = c("MP1", "MP2")
am = nuclearPed(2, children = mp, father = "R1", mother = "R2")
m = marker(am, "R1" = 1:2, "R2"= 1:2)
am = addMarkers(am, m)
res = lik1(vp, mp, pm, am) 
res = lik1(vp[1], mp[1], pm, am) 
}
