% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop2.R
\name{loop2}
\alias{loop2}
\title{Find single candidate list}
\usage{
loop2(vp, mp, sex)
}
\arguments{
\item{vp}{Character vector with names of victims}

\item{sex}{list of sex, one component for victims, 
one for missing persons}

\item{pm}{Character vector with names of missing persons}
}
\value{
Matrix of candidate moves
}
\description{
Find single candidate list
}
\examples{
vp = c("V1", "V2")
mp = c("MP1", "MP2")
mat1 = rbind(c("V1", 1),c("V2", 2))
dimnames(mat1) = list(vp, c("id", "sex"))
mat2 = rbind(c("MP1", 1),c("MP2", 2))
dimnames(mat2) = list(mp, c("id", "sex"))
sex = list(pmsex = mat1, amsex = mat2)
loop2(c("V1", "V2"), c("MP1", "MP2"), sex)


}
