% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopPair.R
\name{loopPair}
\alias{loopPair}
\title{Generate list of candidate pairs to move}
\usage{
loopPair(from1, to1, pm, am)
}
\arguments{
\item{from1}{Character vector with names of victims}

\item{to1}{Character vector with names of missing persons}

\item{pm}{A list of singeltons}

\item{am}{A list of pedigrees}
}
\value{
A data frame with eight columns. The first two describe a candidate pair
to bemoved. The next two a destination pair. Then follows sex information.
}
\description{
All pairs of victims and pairs of missing persons
that are possible accounting for sex are generated.
}
\details{
The output is ordered starting from the first column.
NULL is returned if there are no possible moves.
}
\examples{
from1 = paste("V", 1:3, sep = "")
pm = singletonList(3, sex = c(1, 1, 2))
to1 = paste("MP", 1:4, sep = "")
am = nuclearPed(4, children = to1, sex = c(1, 1, 2, 2))
res = loopPair(from1, to1, pm, am)

}
