% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRdiscrete.R
\name{LRDiscrete}
\alias{LRDiscrete}
\title{LR for discrete property in DVI problems}
\usage{
LRDiscrete(a, x, y, alpha = 0.5, mu = 0.05)
}
\arguments{
\item{a}{Character vector. An assignment}

\item{x}{List with 1 (has property) or 0 (do not have property) for victims}

\item{y}{List with 1 (has property) or 0 (do not have property) for missing persons}

\item{alpha}{Double. P(Y=1)}

\item{mu}{Double. P(X = 1 | Y = 0) = P(X = 0 | Y = 1)}
}
\value{
LR
}
\description{
Based
}
\details{
The LR
}
\examples{
# Example 1
miss = c('*', example2$missing)
lst = list(V1 = miss, V2 = miss, V3 = miss)
tab = expand.grid.nodup2(lst)
nV = dim(tab)[2]
a = tab[1, 1:nV]
a[1:nV] = '*'
x = list(V1 = 1, V2 = 0, V3 = 0)
y = list(M1 = 1, M2 = 0, M3 = 0)
LRDiscrete(a, x, y, alpha = 0.1, mu = 0.01)
na = dim(tab)[1]
foo1 = rep(NA, na)
for (i in 1:na)
foo1[i] = LRDiscrete(tab[i,1:nV], x, y, alpha = 0.05, mu = 0.01)
res = data.frame(tab, foo1)
res = res[order(res$foo1, decreasing = T),]
head(res)

}
\seealso{
\code{\link[=expand.grid.nodup2]{expand.grid.nodup2()}}
}
