% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncomb.R
\name{ncomb}
\alias{ncomb}
\title{The total a priori number of assignments in a DVI problem}
\usage{
ncomb(nVfemales, nMPfemales, nVmales, nMPmales)
}
\arguments{
\item{nVfemales}{Integer. The number of female victims.}

\item{nMPfemales}{Integer. The number of female missing persons.}

\item{nVmales}{Integer. The number of male victims.}

\item{nMPmales}{Integer. The number of male missing persons.}
}
\value{
The total number of possible assignments.
}
\description{
Calculates the total number of assignments, i.e., the different ways the victims can be paired
with the missing persons while respecting sex. Note that in practice, the number of assignments
considered in a joint DVI analysis is usually much smaller than this, due to reductions in
preliminary steps. See \code{\link[=gridSize]{gridSize()}} for a more refined version.
}
\examples{

# Example
m1 = ncomb(5,5,5,5)

# Example: 3 male victims; 2 male missing persons.
# The number of a priori possible assignments is
m1 = ncomb(0,0,3,2) # 13


# Compare with the complete list of assignments
m2 = expand.grid.nodup(list(V1 = c("*", "M1", "M2"),
                            V2 = c("*", "M1", "M2"),
                            V3 = c("*", "M1", "M2")))
stopifnot(m1 == nrow(m2))

}
\seealso{
\code{\link[=gridSize]{gridSize()}}
}
