% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneStep.R
\name{oneStep}
\alias{oneStep}
\title{Forward DVI selection, one at a time}
\usage{
oneStep(from, to, ids.from, ids.to, LRlimit = 1, lik0 = NULL,
  maxid = NULL)
}
\arguments{
\item{from}{A list of singeltons}

\item{to}{A list of pedigrees}

\item{ids.from}{Character vector with names of victims}

\item{ids.to}{Character vector with names of missing persons}

\item{LRlimit}{Double. Threshold for LR}

\item{lik0}{Double.}
}
\value{
output
}
\description{
Forward DVI selection, one at a time
}
\examples{
ids.from = c("V1", "V2")
from = singletonList(2, ids = ids.from)
m = marker(from[[1]], alleles = 1:2, "V1" = 2)
from[[1]] = addMarkers(from[[1]], m)
m = marker(from[[2]], alleles = 1:2, "V2" = 1:2)
from[[2]] = addMarkers(from[[2]], m)
ids.to = c("MP1", "MP2")
to = nuclearPed(2, children = ids.to, father = "R1", mother = "R2")
m = marker(to, "R1" = 1, "R2"= 1:2, alleles = 1:2)
to = addMarkers(to, m)
plotPedList(list(from, to), marker = 1)
lik0 = prod(LR(list(from, to), 1)$likelihoodsPerSystem)
limit = -1
one = oneStep(from, to, ids.from, ids.to, LRlimit = limit, lik0 = lik0)
}
