% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDVI.R
\name{plotSolution}
\alias{plotSolution}
\title{Plot DVI solution}
\usage{
plotSolution(dvi, assignment, k = 1, ...)
}
\arguments{
\item{dvi}{A dviData object}

\item{assignment}{A named character of the format \code{c(victim = missing, ...)},
or a data frame produced by \code{\link[=jointDVI]{jointDVI()}}.}

\item{k}{An integer; the row number when \code{assignment} is a data frame.}

\item{...}{Parameters passed on to \code{\link[=plotDVI]{plotDVI()}}.}
}
\value{
NULL.
}
\description{
Plot DVI solution
}
\examples{

res = jointDVI(example2, verbose = FALSE)

plotSolution(example2, res)
plotSolution(example2, res, marker = 1)

# Non-optimal solutions
plotSolution(example2, res, k = 2, pm = FALSE)
plotSolution(example2, res, k = 2, cex = 1.3)

}
