% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUndisputed.R
\name{plotUndisputed}
\alias{plotUndisputed}
\title{Plot undisputed identifications}
\usage{
plotUndisputed(dvi, undisputed, format = "[S]=[M]", ...)
}
\arguments{
\item{dvi}{A \code{dviData} object, typically created with \code{\link[=dviData]{dviData()}}.}

\item{undisputed}{A data frame containing the undisputed matches, typically
the entry \code{undisputed} in output from \code{\link[=findUndisputed]{findUndisputed()}}
(only three first columns used).}

\item{format}{See \code{\link[=plotSolution]{plotSolution()}}.}

\item{...}{Further parameters to be passed on to \code{\link[pedtools:plot.ped]{pedtools::plot.ped()}},
e.g., \code{marker}, \code{cex}, \code{cex.main}, \code{symbolsize}.}
}
\description{
Plot undisputed identifications
}
\examples{

# Example 
res = findUndisputed(example2, threshold = 2, relax = TRUE, verbose = FALSE)
u = res$undisputed
plotUndisputed(example2, u, marker = 1)

}
