% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\title{remove irelevant victims and missing persons}
\usage{
reduce(pm, am, vp, mp, likNULL = 1, limit = 0)
}
\arguments{
\item{pm}{A list of singeltons}

\item{am}{A list of pedigrees}

\item{vp}{Character vector with names of victims}

\item{mp}{Character vector with names of missing persons}

\item{limit}{Double}

\item{likNull}{Double}
}
\description{
remove irelevant victims and missing persons
}
\examples{
pm1 = singleton("V1")
m = marker(pm1, "V1" = 1, alleles = 1:2)
pm1 = addMarkers(pm1, m)
pm2 = singleton("V2")
m = marker(pm2, "V2" = 2, alleles = 1:2)
pm2 = addMarkers(pm2, m)
p = list(pm1, pm2)
vp = c("V1", "V2", "V3", "V4")
am = nuclearPed(1, father = "MP1", mother = "MP2", children= "R1")
m = marker(am, "R1" = 1, alleles = 1:2)
am = addMarkers(am, m)
mp =c("MP1", "MP2")
reduce(pm, am, vp, mp)

}
