% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceFrom.R
\name{reduceFrom}
\alias{reduceFrom}
\title{Reduce list of from (PM) samples}
\usage{
reduceFrom(from, ids.from, thresholds = c(MZ = 0.99, PO = 0.99))
}
\arguments{
\item{from}{A list of singeltons}

\item{ids.from}{Character vector with names of PM-samples}
}
\description{
Reduce list of from (PM) samples
}
\examples{
library(forrel)
data(dvi.nfi.mut)
dvi.nfi = dvi.nfi.mut
from = dvi.nfi$pm
from = from[c(1:5, c(3,3,4,5))] #copyies
from[[6]]$ID = "V3.c1"
from[[7]]$ID = "V3.c2"
from[[8]]$ID = "V4.c"
from[[9]]$ID = "V5.c"
ids.from = c(dvi.nfi$vict, "V3.c1", "V3.c2", "V4.c", "V5.c")
red = reduceFrom(from,ids.from)
Merge 1, paste navn, fjerne
from = from[-(1:length(from))[ids.from == red[1,2]]]
}
