% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVictims.R
\name{setVictims}
\alias{setVictims}
\alias{addVictims}
\title{Set or add victim samples}
\usage{
setVictims(dvi, vics)

addVictims(dvi, vics)
}
\arguments{
\item{dvi}{A \code{\link[=dviData]{dviData()}} object.}

\item{vics}{A \code{singleton} object or a list of such.}
}
\value{
A modified \code{\link[=dviData]{dviData()}} object.
}
\description{
Convenience functions for replacing or adding victim samples in an existing \code{\link[=dviData]{dviData()}}
object. These functions should generally be preferred over direct manipulation of the
\code{pm} component, as they ensure consistency and recompute the internal \code{pairings}
object.
}
\details{
\code{setVictims()} replaces all existing victim samples, while \code{addVictims()} retains the
existing samples and appends new ones.
}
\examples{
# Add a victim
addVictims(example1, singleton("V4"))

# Replace the PM data
setVictims(example1, singleton("V4"))

}
