% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSearch.R
\name{singleSearch}
\alias{singleSearch}
\title{Single-search LR matrix}
\usage{
singleSearch(
  pm,
  am,
  missing,
  moves = NULL,
  limit = 0.1,
  nkeep = NULL,
  check = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{pm}{A list of singletons, the victims.}

\item{am}{A list of pedigrees. The reference families.}

\item{missing}{A character vector with names of missing persons.}

\item{moves}{A list with possible assignments.}

\item{limit}{A positive number, the lower threshold for LR.}

\item{nkeep}{An integer. No of moves to keep, all if `NULL`.}

\item{check}{A logical, indicating if the input data should be checked for
consistency.}

\item{verbose}{A logical.}
}
\value{
A list with moves and log likelihoods.
}
\description{
For a given DVI problem, compute the LR matrix consisting of individual
likelihood ratios \eqn{LR_{i,j}} comparing the assignment \eqn{V_i = M_j} to
the null. The output may be reduced by specifying arguments `limit` or
`nkeep`.
}
\examples{

pm = example1$pm
am = example1$am
missing = example1$missing

singleSearch(pm, am, missing)

}
