% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletonList.R
\name{singletonList}
\alias{singletonList}
\title{Creates a list of singletons}
\usage{
singletonList(n = 2, ids = paste("V", 1:n, sep = ""),
  famid = paste("F", 1:n, sep = ""), sex = rep(1, n), als = 1:2,
  freq = c(0.5, 0.5), amat = matrix(sample(als, 2 * n, rep = TRUE),
  ncol = 2), seed = NULL)
}
\arguments{
\item{n}{integer Noof singletons}

\item{ids}{Character vector. Names of victims}

\item{famid}{Character vector. Names of families}

\item{sex}{Integer vector}

\item{als}{Alleles}

\item{freq}{Allele frequencies}

\item{amat}{Matrix A nx2 matrix with alleles sampled from and 2}

\item{seed}{Integer}
}
\value{
A list of singletons
}
\description{
Wrapper for `singleton`` function of pedtools. Creates a list of singletons 
with default individual and family ids and alleles sampled from a SNP marker with equally frequent alleles
1 and 2.
}
\examples{
singletonList()
}
