\name{specials}
\alias{noSpecial}
\alias{tikzSpecial}
\title{
  DVI specials handlers
}
\description{
  These objects represent different handlers for processing
  special operations within DVI output.

  The \code{noSpecial} handler does nothing.

  The \code{tikzSpecial} handler processes specials that 
  represent TikZ pictures within a LaTeX document.
}
\usage{
noSpecial
tikzSpecial
}
\author{
  Paul Murrell
}
\seealso{
  \code{\link{grid.tikz}}
}
\keyword{ dplot }
