% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviData.R
\name{subsetDVI}
\alias{subsetDVI}
\title{Extract a subset of a DVI dataset}
\usage{
subsetDVI(dvi, pm = NULL, am = NULL, missing = NULL)
}
\arguments{
\item{dvi}{A \code{\link[=dviData]{dviData()}} object}

\item{pm}{A vector with names or indices of victim samples to include. By
default, all are included.}

\item{am}{A vector with names or indices of AM components include. By
default, all relevant components are included.}

\item{missing}{A vector with names or indices of missing persons to include. By
default, all relevant missing persons are included.}
}
\value{
A \code{dviData} object.
}
\description{
Extract a subset of a DVI dataset
}
\examples{

subsetDVI(example2, pm = 1:2) |> plotDVI()
subsetDVI(example2, pm = "V1", am = 1) |> plotDVI()
subsetDVI(example2, missing = "M3") |> plotDVI()

}
