% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symmetricSibs}
\alias{symmetricSibs}
\title{Dataset: Symmetry examples}
\format{
A \code{dviData} object with the following content:
\itemize{
\item \code{pm}: A list of 5 singletons (male victims).
\item \code{am}: A list of 3 pedigrees
\item \code{missing}: A character vector of length 5, naming the missing persons.
}
}
\usage{
symmetricSibs
}
\description{
A toy DVI dataset illustrating various forms of undecidability due to
symmetries in the solutions.
}
\examples{

symmetricSibs

}
\keyword{datasets}
