\name{texEngine}
\alias{texEngine}
\alias{luatexEngine}
\title{
  TeX Engines
}
\description{
  These objects represent different TeX engines that can
  be selected for processing LaTeX code to produce DVI output.
}
\usage{
texEngine
luatexEngine
}
\author{
  Paul Murrell
}
\seealso{
  \code{\link{grid.latex}}
}
\keyword{ dplot }
