% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFamDVI.R
\name{writeFamDVI}
\alias{writeFamDVI}
\title{Export DVI data to Familias}
\usage{
writeFamDVI(
  x,
  famfile = "dvi.fam",
  params = NULL,
  dbOnly = FALSE,
  openFam = FALSE,
  FamiliasPath = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{dviData} object.}

\item{famfile}{The name or path to the output file to be written. The extension ".fam"
is added if missing.}

\item{params}{A list of Familias parameters controlling database settings, dropout and
other options. See Details.}

\item{dbOnly}{A logical. If TRUE, no pedigree information is included; only the
frequency database.}

\item{openFam}{A logical. If TRUE, an attempt is made to open the produced \code{.fam} file
in an external Familias session. Only available on Windows systems with a working
Familias installation.}

\item{FamiliasPath}{The path to the Familias executable. If empty, the following are
tried in order: "Familias3.exe", "C:/Program Files (x86)/Familias3/Familias3.exe".}

\item{verbose}{A logical, by default TRUE.}
}
\value{
The file name is returned invisibly.
}
\description{
Convenience wrapper for \code{\link[pedFamilias:writeFam]{pedFamilias::writeFam()}} with \code{dvi} parameter set to TRUE.
}
\examples{
writeFamDVI(planecrash, tempfile(fileext = ".fam"))

}
\seealso{
\code{\link[=familias2dvir]{familias2dvir()}}, \code{\link[pedFamilias:writeFam]{pedFamilias::writeFam()}}
}
