% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amDrivenDVI.R
\name{amDrivenDVI}
\alias{amDrivenDVI}
\title{AM driven DVI}
\usage{
amDrivenDVI(dvi, threshold = 10000, threshold2 = 10^3, verbose = TRUE)
}
\arguments{
\item{dvi}{A \code{dviData} object.}

\item{threshold}{LR threshold for 'certain' match}

\item{threshold2}{LR threshold for 'probable' match}

\item{verbose}{TRUE}
}
\value{
A list of \code{dviReduced} and \code{summary}.
}
\description{
Work-in-progress. Note: This function assumes that undisputed identifications
have been removed.
}
\examples{
amDrivenDVI(example2)

# Bigger example: Undisputed first
u = findUndisputed(planecrash)
u$summary

# AM-driven analysis of the remaining
amDrivenDVI(u$dviReduced, threshold2 = 500)

}
