% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amDrivenDVI.R
\name{amDrivenDVI}
\alias{amDrivenDVI}
\title{AM driven DVI}
\usage{
amDrivenDVI(
  dvi,
  simple = TRUE,
  thresholdProbable = 10^3,
  LRmatrix = NULL,
  ignoreSex = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dvi}{A \code{dviData} object.}

\item{simple}{A logical (only TRUE for now)}

\item{thresholdProbable}{LR treshold}

\item{LRmatrix}{NULL}

\item{ignoreSex}{FALSE}

\item{verbose}{TRUE}
}
\value{
A list of \code{dviReduced} and \code{summary}.
}
\description{
Work-in-progress. Note: This function assumes that undisputed identifications
have been removed.
}
\examples{
amDrivenDVI(example2)

# Bigger example: Undisputed first
u = findUndisputed(planecrash)
u$summary
amDrivenDVI(u$dviReduced)

}
