% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkInput}
\alias{checkInput}
\title{Checks input}
\usage{
checkInput(from, to, ids.from, ids.to, moves = NULL)
}
\arguments{
\item{from}{A list of singeltons}

\item{to}{A list of pedigrees}

\item{ids.from}{Character vector with names of victims}

\item{ids.to}{Character vector with names of missing persons}

\item{moves}{A data frame with columns giving the possible moves}
}
\value{
List with two elements. First is `NULL` if no error is detected
and otherwise an error text.The other is the NULL likelihood
}
\description{
Checks input
}
\examples{
checkInput(1, 2, 3, 4)

}
