% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviCompare.R
\name{dviCompare}
\alias{dviCompare}
\title{Compare DVI approaches}
\usage{
dviCompare(
  pm,
  am,
  MPs,
  refs,
  true,
  db,
  Nsim,
  methods = 1:3,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{pm}{PM data: List of singletons}

\item{am}{AM data: A ped object or list of such.}

\item{MPs}{Character vector with names of the missing persons.}

\item{refs}{Character vector with names of the reference individuals.}

\item{true}{A character of the same length as `pm`, with the true solution,
e.g., `true = c("MP2", "*", "MP3)` if the truth is V1 = MP2 and V3 = MP3.}

\item{db}{A frequency database, e.g., forrel::NorwegianFrequencies[1:10].}

\item{Nsim}{A positive integer.}

\item{methods}{A subset of the numbers 1,2,3.}

\item{seed}{A seed for the random number generator, or NULL.}

\item{verbose}{A logical.}
}
\value{
A list of solution frequencies for each method, and a vector of true
  positive rates for each method.
}
\description{
The following methods are compared: 1 = sequential (simple); 2 = sequential
(with updates); 3 = joint.
}
\examples{

pm = example1$pm
am = example1$am
MPs = example1$MPs
refs = "R1"

db = forrel::NorwegianFrequencies[1:10]

# True solution
true = c("MP1", "MP2", "MP3")

# Run comparison
dviCompare(pm, am, MPs, refs, 
           true = true, db = db, Nsim = 2, seed = 1234)

}
