% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviData.R
\name{dviData}
\alias{dviData}
\title{DVI data}
\usage{
dviData(pm, am, missing, generatePairings = TRUE)
}
\arguments{
\item{pm}{A list of singletons: The victim samples.}

\item{am}{A list of pedigrees: The reference families.}

\item{missing}{A character vector with names of missing persons.}

\item{generatePairings}{A logical. If TRUE (default) a list of sex-compatible
pairings is included as part of the output.}
}
\value{
An object of class \code{dviData}, which is basically a list of \code{pm},
\code{am}, \code{missing} and \code{pairings}.
}
\description{
DVI data
}
\examples{
dviData(pm = singleton("V1"), am = nuclearPed(1), missing = "3")

}
