% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviSearch.R
\name{dviSearch}
\alias{dviSearch}
\title{Search for DVI solution}
\usage{
dviSearch(from, to, ids.from, ids.to, limit = 1, nbest = 2,
  extend = TRUE, pl = FALSE)
}
\arguments{
\item{from}{A list of singeltons}

\item{to}{A list of pedigrees}

\item{ids.from}{Character vector with names of victims}

\item{ids.to}{Character vector with names of missing persons}

\item{limit}{Double. Threshold for LR}

\item{nbest}{Integer If not `NULL`, search is restricted to the `nbest`
marginal solutions for each victim}

\item{extend}{Logical. If `TRUE`any number of victims need not be among 
the missing persons}

\item{pl}{Logical. If `TRUE`, a plot of the best solution will be made. A plot window should 
be prepared and sized in advance}
}
\value{
A data frame with moves sorted according to LR. Possibly a plot of the best solution.
}
\description{
A solution is obtain by finding the most likely among marginal 
possible moves
}
\details{
The search can be restricted using the `nbest` variable.
}
\examples{
\dontrun{
data(dvi.nfi)
data(dvi.nfi.mut)
dvi.nfi = dvi.nfi.mut
from = dvi.nfi$pm
to = dvi.nfi$am
ids.from = dvi.nfi$vict
ids.to = dvi.nfi$miss
limit = 1e+100
extend = TRUE
nbest = 1
pl = TRUE
foo = dviSearch(from, to, ids.from,ids.to, limit = limit, 
nbest = nbest, extend = extend, pl = pl)
}
}
