% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviSolve.R
\name{dviSolve}
\alias{dviSolve}
\title{A complete pipeline for solving a DVI case}
\usage{
dviSolve(
  dvi,
  threshold = 10000,
  threshold2 = max(1, threshold/10),
  maxIncomp = 2,
  ignoreSex = FALSE,
  limit = 0,
  detailedOutput = FALSE,
  verbose = TRUE,
  debug = FALSE
)
}
\arguments{
\item{dvi}{A \code{dviData} object.}

\item{threshold}{LR threshold for 'significant' match.}

\item{threshold2}{LR threshold for 'probable' match. By default set to
\code{threshold/10}.}

\item{maxIncomp}{An integer passed onto \code{\link[=findExcluded]{findExcluded()}}. A pairing is
excluded if the number of incompatible markers exceeds this.}

\item{ignoreSex}{A logical, by default FALSE.}

\item{limit}{A number passed onto \code{\link[=findUndisputed]{findUndisputed()}}; only pairwise LR values
above this are considered.}

\item{detailedOutput}{A logical, by default FALSE. See Details.}

\item{verbose, debug}{Logicals.}
}
\value{
A list of data frames \code{AM} and \code{PM}. If \code{detailedOutput} is TRUE, the
LR matrix and exclusion matrix from the first iteration are also included.
}
\description{
This wraps several other functions into a complete pipeline for solving a DVI
case.
}
\examples{
dviSolve(example2)
dviSolve(example2, threshold = 5, detailedOutput = TRUE, verbose = FALSE)

}
