% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{DVI dataset: Generational trio}
\format{
A list of 3 elements:
\itemize{
\item \code{pm}: A list of 3 singletons (victims).
\item \code{am}: A pedigree with three missing persons and one typed reference
individual.
\item \code{missing}: A vector containing the names of the missing persons.
}
}
\usage{
example1
}
\description{
A proof-of-concept dataset involving three missing members (child, father,
grandfather) of a single family. With the given data, stepwise victim
identification fails to find the correct solution, while joint identification
succeeds.
}
\keyword{datasets}
