% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{example1}
\format{
A list of 3 elements:

  * `pm`: A list of 3 singletons (victims).

  * `am`: A pedigree with three missing persons and one typed reference
  individual.

  * `MPs`: A vector containing the names of the missing persons.
}
\usage{
example1
}
\description{
A proof-of-concept dataset showing that showing that stepwise victim
identification may fail
}
\keyword{datasets}
