% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exclusionExample}
\alias{exclusionExample}
\title{Data. exclusionExample}
\format{
A list of 3 elements:
\itemize{
\item \code{pm}: A list of 16 singletons (male victims).
\item \code{am}: A list of 15 pedigrees, each with one missing person
\item \code{missing}: A vector containing the names of the 15 missing persons.
}
}
\usage{
exclusionExample
}
\description{
This data is based on a real case, but pedigrees have been changed and
marker data simulated to preserve anonymity.
}
\examples{

pm = exclusionExample$pm
am = exclusionExample$am
missing = exclusionExample$missing
summariseDVI(pm, am , missing)
}
\keyword{datasets}
