% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise497.R
\name{exercise497}
\alias{exercise497}
\title{Code. Exercise 4.9.7 in Kling et al. (2021)}
\usage{
exercise497(
  pm,
  am,
  missing,
  nsim = 2,
  seed = NULL,
  simRef = TRUE,
  disableMutations = FALSE,
  undisputed = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pm}{A list of singletons.}

\item{am}{A list of pedigrees.}

\item{missing}{Character vector with names of missing persons.}

\item{nsim}{Number of simulations.}

\item{seed}{Integer.}

\item{simRef}{Logical. if TRUE, references are also simulated.}

\item{disableMutations}{Logical, see [jointDVI()].}

\item{undisputed}{Logical, see [jointDVI()].}

\item{verbose}{A logical}
}
\value{
A list with two elements, the first the fraction of 'correct' solutions, the second
a matrix with first line from [jointDVI()]
}
\description{
This is a DVI case with 3 female victims and 3 missing females
in three reference families. There are 23 markers with equal mutation rate 0.001.
Data are simulated from the solution V1 MP1, V2 = MP2, V3 = MP3 and the purpose is
to check fraction of times the 'correct' solutions is obtained.
}
\examples{
pm = dataExercise497$pm
am = dataExercise497$am
missing = dataExercise497$missing

exercise497(pm, am, missing, nsim = 2, seed = 17, verbose = TRUE)
\donttest{
exercise497(pm, am, missing, nsim = 1, seed = 17, verbose = TRUE, fixUndisputed = TRUE)
}
}
\seealso{
[jointDVI()]
}
