% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.nodup.R
\name{expand.grid.nodup}
\alias{expand.grid.nodup}
\title{Generate list of assignments}
\usage{
expand.grid.nodup(lst)
}
\arguments{
\item{lst}{List or 0-1 matrix describing possible moves.}
}
\value{
List of data frames.
}
\description{
Based on a list or a matrix describiing possible moves of victims to missing
persons, all unique possibilities are generated.
}
\examples{

# Example 1
moves = list(V1 = c("V1", "MP1"), V2 = c("V2", "MP2"))
expand.grid.nodup(moves)
expand.grid.nodup(moves[1])

# Example 2
expand.grid.nodup(list(c(1,2,4), c(1,3), 2))

# Example 3
expand.grid.nodup(list(c(12, 1), 2))

\dontrun{
lst = rep(list(1:12), 5)
gr = expand.grid.nodup(lst)   # takes a few seconds
length(gr)   # 95040
}

}
\author{
Magnus Dehli Vigeland
}
