% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.nodup.R
\name{expand.grid.nodup}
\alias{expand.grid.nodup}
\title{Generate list of assignments}
\usage{
expand.grid.nodup(lst)
}
\arguments{
\item{lst}{List or 0-1 matrix describing possible moves}
}
\value{
List with possible moves
}
\description{
Based on a list or a matrix describiing possible
moves of victims to missing persons, all
unique possibilies are generated.
}
\examples{
expand.grid.nodup(list(c(1,2,4), c(1,3), 2))
lst = list(c("MP12", "MP1"), c("MP2"))
lst = list(c(12, 1), 2)
expand.grid.nodup(lst)
\dontrun{
lst = rep(list(1:12), 5)
gr = expand.grid.nodup(lst)   # tar noen sekunder
length(gr)   # 95040
}

}
