% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.nodup.R
\name{expand.grid.nodup}
\alias{expand.grid.nodup}
\title{Combinations without duplications}
\usage{
expand.grid.nodup(lst)
}
\arguments{
\item{lst}{A list of vectors.}
}
\value{
A data frame.
}
\description{
This is a simple extension of \code{\link[=expand.grid]{expand.grid()}} which removes all combinations
with repeated elements.
}
\examples{

lst = list(1:2, 1:2)

# Compare
expand.grid.nodup(lst)
expand.grid(lst)


}
\seealso{
\code{\link[=expand.grid]{expand.grid()}}
}
