% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusionMatrix.R
\name{findExcluded}
\alias{findExcluded}
\title{Excluded pairings}
\usage{
findExcluded(
  dvi,
  pairings = NULL,
  ignoreSex = FALSE,
  maxIncomp = 2,
  verbose = TRUE
)
}
\arguments{
\item{dvi}{A \code{\link[=dviData]{dviData()}} object.}

\item{pairings}{A list of possible pairings for each victim. If NULL, all
sex-consistent pairings are used, unless if \code{ignoreSex} is set to TRUE (in
which case all pairings are considered).}

\item{ignoreSex}{A logical, by default FALSE.}

\item{maxIncomp}{An integer. A pairing is excluded if the number of
incompatible markers exceeds this.}

\item{verbose}{A logical. Default: TRUE.}
}
\value{
A list with the following entries:
\itemize{
\item \code{exclusionMatrix}: A matrix showing the number of inconsistencies for
each pair (or NA if the pairing was not considered)
\item \code{pmNomatch}: A vector with the names of victims (PM singletons) who are
excluded against all missing persons.
\item \code{missNomatch}: A vector with the missing persons who are excluded against
all victims.
\item \code{dviReduced}: A reduced version of \code{dvi}, where excluded victims/missing
persons are removed.
\item \code{pairings} A list of non-excluded pairings.
}
}
\description{
Excluded pairings
}
\examples{
findExcluded(icmp)

}
