% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUndisputed.R
\name{findUndisputed}
\alias{findUndisputed}
\title{Undisputed identifications in DVI problems}
\usage{
findUndisputed(
  pm,
  am,
  missing,
  threshold = 10000,
  limit = 0,
  check = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{pm}{PM data: List of singletons.}

\item{am}{AM data: A `ped` object or list of such.}

\item{missing}{Character vector with names of the missing persons.}

\item{threshold}{A non-negative number. If no single-search LR values exceed
this, the iteration stops.}

\item{limit}{A positive number. Only single-search LR values above this are
considered.}

\item{check}{A logical, indicating if the input data should be checked for
consistency.}

\item{verbose}{A logical.}
}
\value{
A list with the following entries:

  * `undisputed`: A list of undisputed matches and the corresponding LR
  values.

  * `pmReduced`: Same as `pm`, but with the undisputed victims removed.

  * `amReduced`: Same as `am`, but with the data from undisputed victims
  inserted.

  * `missingReduced`: Same as `missing`, but without the undisputed
  identified missing persons.

  * `moves`, `LR.table`, `LRmatrix`: Output from `singleSearch()` applied to
  the reduced problem.
}
\description{
This function uses the single-search LR values to find "undisputed" matches
between victims and missing individuals. An identification \eqn{V_i = M_j} is
called undisputed if the corresponding likelihood ratio \eqn{LR_{i,j}}
exceeds the given threshold, while all other values involving \eqn{v_i} or
\eqn{M_j} are below 1.
}
\examples{

pm = planecrash$pm
am = planecrash$am
missing = planecrash$missing

findUndisputed(pm, am, missing, threshold = 1e4)

}
