% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUndisputed.R
\name{findUndisputed}
\alias{findUndisputed}
\title{Undisputed identifications in DVI problems}
\usage{
findUndisputed(
  pm,
  am,
  missing,
  threshold = 10000,
  check = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{pm}{PM data: List of singletons.}

\item{am}{AM data: A `ped` object or list of such.}

\item{missing}{Character vector with names of the missing persons.}

\item{threshold}{A non-negative number. If no single-search LR values exceed this,
the iteration stops.}

\item{check}{A logical, indicating if the input data should be checked for
consistency.}

\item{verbose}{A logical.}
}
\value{
A named character vector.
}
\description{
This function uses the single-search LR values to find "undisputed" matches
between victims and missing individuals. An identification `V_i = M_j` is
called undisputed if the corresponding likelihood ratio `LR_{i,j}` exceeds
the given threshold, while all other values involving `v_i` or `M_j` are
below 1.
}
\examples{

pm = planecrash$pm
am = planecrash$am
missing = planecrash$MPs

findUndisputed(pm, am, missing, threshold = 1e4)

}
