% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freqsBlind}
\alias{freqsBlind}
\title{Allele frequencies
A dataset `freqsBlind` containing allele frequencies}
\format{
A list with 198 elements, markers.  The first 27 are for STRs, the remaining for SNps
}
\usage{
freqsBlind
}
\description{
Allele frequencies
A dataset `freqsBlind` containing allele frequencies
}
\keyword{datasets}
