% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Find next candidates}
\usage{
generate(pm, am, vp, mp, one = TRUE)
}
\arguments{
\item{pm}{A list of singeltons}

\item{am}{A list of pedigrees}

\item{vp}{Character vector with names of victims}

\item{mp}{Character vector with names of missing persons}

\item{one}{logic If TRUE, only one-step}
}
\value{
list of matrices of one and two-moves
}
\description{
New moves are generated, NULL
if none are possible, for one and two step
}
\examples{
vp = c("V1", "V2")
pm = singletonList(2, ids = vp)
m = marker(pm[[1]], alleles = 1:2, "V1" = 1)
pm[[1]] = addMarkers(pm[[1]], m)
m = marker(pm[[2]], alleles = 1:2, "V2" = 1:2)
pm[[2]] = addMarkers(pm[[2]], m)
mp = c("MP1", "MP2")
am = nuclearPed(2, children = mp, father = "R1", mother = "R2")
m = marker(am, "R1" = 1:2, "R2"= 1:2)
am = addMarkers(am, m)
# The thre possibilities: a move of two, one or none:
tab = generate(pm, am, vp, mp)
pm[[1]] = swapSex(pm[[1]], "V1")
tab = generate(pm, am, vp,mp)
pm[[2]] = swapSex(pm[[2]], "V2")
tab = generate(pm, am, vp,mp)
}
