% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMoves.R
\name{generateMoves}
\alias{generateMoves}
\title{Generate list of possible moves}
\usage{
generateMoves(pm, am, missing, ignore.sex = FALSE, expand.grid = FALSE)
}
\arguments{
\item{pm}{A list of singletons.}

\item{am}{A list of pedigrees.}

\item{missing}{Character vector with names of missing persons.}

\item{ignore.sex}{A logical.}

\item{expand.grid}{A logical. If TRUE, [expand.grid.nodup()] is applied to
the result before returning.}
}
\value{
A list of moves.
}
\description{
A sex-consistent list of moves for [jointDVI()] is generated
}
\details{
Identity moves are included.
}
\examples{

library(pedtools)

pm = list(singleton("V1", sex = 1),
            singleton("V2", sex = 2))
missing = paste0("MP", 1:4)
am = list(nuclearPed(children = missing[1:3]),
          nuclearPed(children = missing[4], sex = 2))
generateMoves(pm, am, missing)

}
\seealso{
[jointDVI()]
}
