% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointDVI.R
\name{jointDVI}
\alias{jointDVI}
\alias{global}
\alias{checkDVI}
\title{Joint DVI search}
\usage{
jointDVI(
  pm,
  am,
  missing,
  moves = NULL,
  limit = 0,
  fixUndisputed = TRUE,
  numCores = 1,
  check = TRUE,
  verbose = FALSE
)

checkDVI(pm, am, missing, moves)
}
\arguments{
\item{pm}{A list of singletons.}

\item{am}{A list of pedigrees.}

\item{missing}{Character vector with names of missing persons.}

\item{moves}{List of length equal length of `pm` with possible marginal moves.}

\item{limit}{Double. Only moves with LR above limit are kept.}

\item{numCores}{Integer. The number of cores used in parallelisation. Default: 1.}

\item{check}{A logical, indicating if the input data should be checked for consistency.}

\item{verbose}{A logical.}
}
\value{
A data frame. Each row describes an a priori possible move. The log likelihood, 
the LR with the null hypothesis of no moves in the numerator, and the posterior corresponding 
to a flat prior, i.e., the inverse of the number of assignments.
}
\description{
Victims are given as a list of singletons and references as list of pedigrees.
Based on a specification of moves for each victim, if any, all possible assignments 
are evaluated and solutions ranked according to the likelihood..
}
\examples{

\donttest{
library(forrel)

### Example 1 ###

# Attributes of a single marker
locAttr = list(name = "m", alleles = 1:3, afreq = c(1, 1, 1)/3)

# PM data (victims: 6 males, 1 female)
vics = paste0("V", 1:7)
sex = c(1,1,1,1,1,1,2)
df = data.frame(famid = vics, id = vics, fid = 0, mid = 0, sex = sex,
                m = c("1/1", "2/2", "1/1", "1/1", "2/2", "2/2", "2/2"))
pm = as.ped(df, locusAttributes = locAttr)

# AM data (families)
missing = c("MP1", "MP2", "MP3")
am = nuclearPed(3, father = "R1", mother = "R2", children = missing)
m = marker(am, "R1" = "1/1", "R2" = "1/1", name = "m")
am = setMarkers(am, m, locusAttributes = locAttr)

# Plot
plotPedList(list(pm, am), marker = 1, col = list(red = missing),
            titles = c("PM data", "AM data"))

# Analysis considering all sex-consistent assignments
res1 = jointDVI(pm, am, missing, moves = NULL, limit = -1, verbose = FALSE)

# Quicker alternative: Consider only the three best moves for each victim
moves = generateMoves(pm, am, missing) # generate all sex-consistent assignments
moves2 = marginal(pm, am,  missing, moves, limit = -1, nkeep = 3)
res2 = jointDVI(pm, am, missing, moves2[[1]], limit = -1, verbose = FALSE)

# Further reduction: Only consider victims V1, V3 and V4
# moves2 = moves[c("V1", "V3", "V4")]
# res = jointDVI(pm, am, missing, moves2, limit = -1)

### Example 2 ###
# Checked against; http://familias.name/BookKEP/globalExample2.fam

# Single locus with 10 alleles
loc = list(alleles = 1:10)

# Victims: 3 males, 1 female
vics = paste0("V", 1:4)
sex = c(1, 1, 2, 1)
df = data.frame(famid = vics, id = vics, fid = 0, mid = 0, sex = sex,
                m1 = c("1/1", "1/2", "3/4", "3/4"))
pm = as.ped(df, locusAttributes = loc)

# Reference families: 4 missing persons; 2 genotyped relatives
fam1 = nuclearPed(1, father = "MP1", child = "MP2", mother ="R1")
fam2 = halfSibPed(sex1 = 1, sex2 = 2)
fam2 = relabel(fam2, c("MO2", "R2", "MO3", "MP3", "MP4"))
data = data.frame(m1 = c(R1 = "2/2", R2 = "3/3"))
am = setMarkers(list(fam1, fam2), alleleMatrix = data, locusAttributes = loc)

# Generate sex-consistent moves
missing = c("MP1", "MP2", "MP3", "MP4") # user specified
moves = generateMoves(pm, am, missing)

# Rank according to likelihood
res1 = jointDVI(pm, am, missing, moves = moves, limit = 0, verbose = TRUE)
resmarg = marginal(pm, am, missing, moves = moves, limit = 0,
             verbose = TRUE, nkeep = 2)
res2 = jointDVI(pm, am, missing, moves = resmarg[[1]], limit = 0, verbose = TRUE)

# Victims fam file
x = readFam("http://familias.name/BookKEP/globalExample2.fam")
pm = x$families[1:4]
am = x$families[5:6]
plotPedList(list(pm, am))
res3 = jointDVI(pm, am, missing, moves = moves, limit = 0, verbose = TRUE)

stopifnot(identical(res1, res3))

}

}
\seealso{
`marginal`.
}
