% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik1.R
\name{lik1}
\alias{lik1}
\title{Move individuals, find likelihoods and update}
\usage{
lik1(from, to, ids.from, ids.to)
}
\arguments{
\item{from}{A list of singeltons}

\item{to}{A list of pedigrees}

\item{ids.from}{Character vector with names of victims}

\item{ids.to}{Character vector with names of missing persons}
}
\value{
Likelihood and updated from and to lists
}
\description{
Individuals are moved from pm to am data.
The likelihood is calculated and pm and am
are updated
}
\examples{
ids.from = c("V1", "V2")
from = singletonList(2, ids = ids.from)
m = marker(from[[1]], alleles = 1:2, "V1" = 1:2)
from[[1]] = setMarkers(from[[1]], m)
m = marker(from[[2]], alleles = 1:2, "V2" = 1:2)
from[[2]] = setMarkers(from[[2]], m)
ids.to = c("MP1", "MP2")
to = nuclearPed(2, children = ids.to, father = "R1", mother = "R2")
m = marker(to, "R1" = 1:2, "R2"= 1:2)
to = setMarkers(to, m)
res = lik1(from, to, ids.from, ids.to) 
res = lik1(from, to, ids.from[1], ids.to[1]) 
}
