% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncomb.R
\name{ncomb}
\alias{ncomb}
\title{Finds number of combinations for DVI problem}
\usage{
ncomb(nVfemales, nMPfemales, nVmales, nMPmales)
}
\arguments{
\item{nVfemales}{Integer. Number of female victims.}

\item{nMPfemales}{Integer. Number of missing persons, females.}

\item{nVmales}{Integer. Number of male victims.}

\item{nMPmales}{Integer. Number of missing persons,  males.}
}
\value{
Number of possible moves
}
\description{
The number of victims and missing persons and their sex is given. The number
of possible moves, i,e., the number of ways the victims can be identified
with the missing persons is calculated.
}
\examples{
# There are 4 female victims and four female missing persons
ncomb(4, 4, 0, 0) # = 209
# If there are 3 females victims and 3 missing females
# and 1 male victim and missing
ncomb(3, 3, 1, 1)

\dontrun{
# The above examples can be checked as below
ids.to = c("MP1", "MP2", "MP3", "MP4") # Female victims
moves = list(V1 = c("V1", ids.to), V2 = c("V2", ids.to),
             V3 = c("V3", ids.to), V4 = c("V4", ids.to))
foo = expand.grid.nodup(moves)
length(foo)

ids.to = c("MP1", "MP2", "MP3") # Female victims
moves = list(V1 = c("V1", ids.to), V2 = c("V2", ids.to),
             V3 = c("V3", ids.to), V4= c("V4", "MP4"))
foo = expand.grid.nodup(moves)
length(foo) 

}
}
