% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncomb.R
\name{ncomb}
\alias{ncomb}
\title{Finds number of combinations for DVI problem}
\usage{
ncomb(nVfemales, nMPfemales, nVmales, nMPmales)
}
\arguments{
\item{nVfemales}{Integer. Number of female victims.}

\item{nMPfemales}{Integer. Number of missing persons, females.}

\item{nVmales}{Integer. Number of male victims.}

\item{nMPmales}{Integer. Number of missing persons,  males.}
}
\value{
Number of possible moves
}
\description{
A formula is implemented
}
\examples{
There 3 female victims, 6 missing females, 2 male victims, 6 missing males
ncomb(3,6,2,6) # = 9847 As checked by counting

\dontrun{
moves =list(c(1,3,8,12,5,9,101),
c(7,2,10,11,6,4,102),
c(3,12,1,8,5,9,103),
c(4,7,2,10,11,6,104),
c(5,12,3,8,9,1,105))
foo = expand.grid.nodup(moves)
length(foo)
}
}
