% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneStep.R
\name{oneStep}
\alias{oneStep}
\title{Forward DVI selection, one at a time}
\usage{
oneStep(pm, am, vp, mp, LRlimit = 1, lik0 = NULL, maxid = NULL)
}
\arguments{
\item{pm}{A list of singeltons}

\item{am}{A list of pedigrees}

\item{vp}{Character vector with names of victims}

\item{mp}{Character vector with names of missing persons}

\item{LRlimit}{Double. Threshold for LR}

\item{lik0}{Double.}
}
\value{
output
}
\description{
Forward DVI selection, one at a time
}
\examples{
vp = c("V1", "V2")
pm = singletonList(2, ids = vp)
m = marker(pm[[1]], alleles = 1:2, "V1" = 1:2)
pm[[1]] = addMarkers(pm[[1]], m)
m = marker(pm[[2]], alleles = 1:2, "V2" = 1:2)
pm[[2]] = addMarkers(pm[[2]], m)
mp = c("MP1", "MP2")
am = nuclearPed(2, children = mp, father = "R1", mother = "R2")
m = marker(am, "R1" = 1:2, "R2"= 1:2)
am = addMarkers(am, m)
lik0 = prod(LR(list(pm, am), 1)$likelihoodsPerSystem)
limit = -1
one = oneStep(pm, am, vp, mp, LRlimit = limit, lik0 = lik0)
}
