% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planecrash}
\alias{planecrash}
\title{Data. planecrash}
\format{
A list of 3 elements:

  * `pm`: A list of 8 female singletons (victims).

  * `am`: A list of 5 pedigrees, each with one missing member and one
  genotyped member.

  * `missing`: A vector containing the names of the missing persons.
}
\usage{
planecrash
}
\description{
DVI dataset based on simulated data in Exercise 3.3 in Egeland et al.
"Relationship Inference with Familias and R" (2015).
}
\keyword{datasets}
