% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDVI.R
\name{plotDVI}
\alias{plotDVI}
\title{Plot a DVI problem}
\usage{
plotDVI(
  dvi,
  hatched = typedMembers,
  col = list(red = dvi$missing),
  frames = TRUE,
  titles = c("PM", "AM"),
  widths = NULL,
  dev.height = NULL,
  dev.width = NULL,
  newdev = !is.null(c(dev.height, dev.width)),
  ...
)
}
\arguments{
\item{dvi}{A \code{dviData} object, typically created with \code{\link[=dviData]{dviData()}}.}

\item{hatched}{A character vector of ID labels, or the name of a function. By
default, typed individuals are hatched.}

\item{col}{A list of colour vectors (see \code{\link[pedtools:plot.ped]{pedtools::plot.ped()}}). By default,
missing members of \code{dvi$am} are shown in red.}

\item{frames}{A logical, by default TRUE.}

\item{titles}{A character of length 2.}

\item{widths}{A numeric with relative plot widths.}

\item{dev.height, dev.width}{Plot height and widths in inches. These are
optional, and only relevant if \code{newdev = TRUE}.}

\item{newdev}{A logical indicating if a new plot window should be opened.}

\item{...}{Further parameters to be passed on to \code{\link[pedtools:plot.ped]{pedtools::plot.ped()}}, e.g.,
\code{marker}, \code{cex}, \code{cex.main}, \code{symbolsize}.}
}
\description{
Plot a DVI problem
}
\examples{

plotDVI(example1, marker = 1)

# plotDVI(icmp)

# plotDVI(example2, new = T, frames = FALSE, marker = 1, cex = 1.2)

}
