% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviData.R
\name{plotDVI}
\alias{plotDVI}
\title{Plot a DVI problem}
\usage{
plotDVI(dvi, hatched = typedMembers, col = list(red = dvi$missing), ...)
}
\arguments{
\item{dvi}{A \code{dviData} object, typically created with \code{\link[=dviData]{dviData()}}.}

\item{hatched}{A character vector of ID labels, or the name of a function. By
default, typed individuals are hatched.}

\item{col}{A list of colour vectors (see \code{\link[pedtools:plot.ped]{pedtools::plot.ped()}}). By default,
missing members of \code{dvi$am} are shown in red.}

\item{...}{Further parameters to be passed on to \code{\link[pedtools:plotPedList]{pedtools::plotPedList()}}.
Useful parameters include \code{frames}, \code{marker}, \code{newdev} and \code{titles}.}
}
\description{
Plot a DVI problem
}
\examples{

plotDVI(example1)

# plotDVI(example2, new = T, frames = FALSE, marker = 1, cex = 1.2)

# plotDVI(icmp)

}
