% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUndisputed.R
\name{plotUndisputed}
\alias{plotUndisputed}
\title{Plot undisputed identifications}
\usage{
plotUndisputed(dvi, undisputed, ...)
}
\arguments{
\item{dvi}{A \code{dviData} object.}

\item{undisputed}{A data frame containing the undisputed matches, typically
the entry \code{undisputed} in output from \code{\link[=findUndisputed]{findUndisputed()}} (only three first
columns used).}

\item{...}{Further arguments passed on to \code{\link[=plotSolution]{plotSolution()}}.}
}
\description{
Plot undisputed identifications
}
\examples{

# Example
res = findUndisputed(example2, threshold = 2, verbose = FALSE)
u = res$undisputed
plotUndisputed(example2, u, marker = 1)

}
\seealso{
\code{\link[=findUndisputed]{findUndisputed()}}, \code{\link[=plotSolution]{plotSolution()}}
}
