% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelDVI.R
\name{relabelDVI}
\alias{relabelDVI}
\title{Automatic labelling of a DVI dataset}
\usage{
relabelDVI(
  dvi,
  victimPrefix = NULL,
  familyPrefix = NULL,
  refPrefix = NULL,
  missingPrefix = NULL,
  missingFormat = NULL,
  othersPrefix = NULL
)
}
\arguments{
\item{dvi}{A \code{dviData} object, typically created with \code{\link[=dviData]{dviData()}}.}

\item{victimPrefix}{Prefix used to label PM individuals.}

\item{familyPrefix}{Prefix used to label the AM families.}

\item{refPrefix}{Prefix used to label the reference individuals, i.e., the
typed members of the AM families.}

\item{missingPrefix}{Prefix used to label the missing persons in the AM
families. The word "family" is treated as a special case, where the family
name is used as prefix in each family, e.g., F1-1, F1-2, F2-1, ...}

\item{missingFormat}{A string indicating family-wise labelling of missing
persons, using \verb{[FAM]} an \verb{[IDX]} as place holders for the family index and
the missing person index within the family. See Examples.}

\item{othersPrefix}{Prefix used to label other untyped individuals. Default:
1, 2, ...}
}
\value{
A list with entries "pm", "am" and "missing".
}
\description{
Relabel the families and individuals in a DVI dataset, using automatic
labelling.
}
\details{
By default, the following labelling scheme is applied:
\itemize{
\item Victims (PM data): V1, V2, ...
\item Reference families: F1, F2, ...
\item Reference individuals: R1, R2, ...
\item Missing persons: M1, M2, ...
\item Others: 1, 2, ...
}
}
\examples{

# Builtin dataset `example2`
relabelDVI(example2,
           victimPrefix  = "vic",
           familyPrefix  = "fam",
           refPrefix     = "ref",
           missingPrefix = "mp")

# Family-wise labelling of missing persons
relabelDVI(example2, missingFormat = "M[FAM]-[IDX]")
relabelDVI(example2, missingFormat = "M[IDX] (F[FAM])")
relabelDVI(example2, missingFormat = "fam[FAM].m[IDX]")

}
