% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelDVI.R
\name{relabelDVI}
\alias{relabelDVI}
\title{Automatic labelling of a DVI dataset}
\usage{
relabelDVI(
  pm,
  am,
  missing,
  victimPrefix = "V",
  familyPrefix = "F",
  refPrefix = "R",
  missingPrefix = "M"
)
}
\arguments{
\item{pm}{A list of singletons.}

\item{am}{A list of pedigrees.}

\item{missing}{Character vector with names of missing persons.}

\item{victimPrefix}{Prefix used to label PM individuals.}

\item{familyPrefix}{Prefix used to label the AM families.}

\item{refPrefix}{Prefix used to label the reference individuals, i.e., the
typed members of the AM families.}

\item{missingPrefix}{Prefix used to label the missing persons in the AM
families. The word "family" is treated as a special case, where the family
name is used as prefix in each family.}
}
\value{
A list with entries "pm", "am" and "missing".
}
\description{
Relabel the families and individuals in a DVI dataset, using
}
\examples{

relabel(example1$pm, example1$am, example1$missing)

}
