% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequentialDVI.R
\name{sequentialDVI}
\alias{sequentialDVI}
\title{Sequential DVI search}
\usage{
sequentialDVI(
  dvi,
  updateLR = TRUE,
  threshold = 1,
  check = TRUE,
  verbose = TRUE,
  debug = FALSE
)
}
\arguments{
\item{dvi}{A \code{dviData} object, typically created with \code{\link[=dviData]{dviData()}}.}

\item{updateLR}{A logical. If TRUE, the LR matrix is updated in each
iteration.}

\item{threshold}{A non-negative number. If no pairwise LR values exceed this,
the iteration stops.}

\item{check}{A logical, indicating if the input data should be checked for
consistency.}

\item{verbose}{A logical.}

\item{debug}{A logical. If TRUE, the LR matrix is printed}
}
\value{
A solution to the DVI problem in the form of an assignment vector.
}
\description{
Sequential DVI search
}
\examples{
sequentialDVI(example1, updateLR = FALSE)
sequentialDVI(example1, updateLR = TRUE)

# The output of can be fed into `jointDVI()`:
res = sequentialDVI(example1, updateLR = TRUE)
jointDVI(example1, assignments = res)

}
