% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletonList.R
\name{singletonList}
\alias{singletonList}
\title{Creates a list of singlelots}
\usage{
singletonList(n = 2, ids = paste("V", 1:n, sep = ""),
  famid = paste("F", 1:n, sep = ""), sex = rep(1, n))
}
\arguments{
\item{n}{integer Noof singletons}

\item{ids}{Character vector. Names of victims}

\item{famid}{Character vector. Names of families}

\item{sex}{Integer vector}
}
\value{
Alist of singletons
}
\description{
Wraper for singleton function of pedtools
}
\examples{
singletonList()
}
